Email_Tasks/requirements.txt
plyer
openai
python-dotenv